Rename all Alter, Process and Preprocess socialbase classes
This commit removes the unnecessary prefix from all classes. That
function is already fulfilled by PHP namespaces.

There were some inconsistensies in filenames and classnames
capitalisation. These inconsistensies have been left for what they
are because existing sites might be using them.

This should provide a backwards compatible solution to using the
proper classnames and all prefixed classes have been marked
deprecated with a reference to their replacement unprefix class.

No tests should need changing.

Also fixes a few coding style issues in the files.
