+++ themes/socialbase/src/Plugin/Preprocess/SocialBaseDetails.php
@@ -2,28 +2,11 @@
 
 namespace Drupal\socialbase\Plugin\Preprocess;
 
-use Drupal\bootstrap\Plugin\Preprocess\PreprocessBase;
-
 /**
  * Pre-processes variables for the "details" theme hook.
  *
  * @ingroup plugins_preprocess
- *
- * @BootstrapPreprocess("details")
+ * @deprecated
+ * @see \Drupal\socialbase\Plugin\Preprocess\Details
  */
-class SocialBaseDetails extends PreprocessBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  public function preprocess(array &$variables, $hook, array $info) {
-    parent::preprocess($variables, $hook, $info);
-
-    // Do not display the details title in file upload widget.
-    if (isset($variables['element']['#theme']) && $variables['element']['#theme'] == 'file_widget_multiple') {
-      $variables['title'] = FALSE;
-    }
-
-  }
-
-}
+class SocialBaseDetails extends Details {}
