+++ themes/socialbase/src/Plugin/Preprocess/SocialBaseSelect.php
@@ -2,27 +2,11 @@
 
 namespace Drupal\socialbase\Plugin\Preprocess;
 
-use Drupal\bootstrap\Plugin\Preprocess\Select;
-
 /**
  * Pre-processes variables for the "select" theme hook.
  *
  * @ingroup plugins_preprocess
- *
- * @BootstrapPreprocess("select")
+ * @deprecated
+ * @see \Drupal\socialbase\Plugin\Preprocess\Select
  */
-class SocialBaseSelect extends Select {
-
-  /**
-   * {@inheritdoc}
-   */
-  public function preprocess(array &$variables, $hook, array $info) {
-    parent::preprocess($variables, $hook, $info);
-
-    if (isset($variables['element']['#multiple']) && $variables['element']['#multiple'] == TRUE) {
-      $variables['multiselect'] = TRUE;
-    }
-
-  }
-
-}
+class SocialBaseSelect extends Select {}
