+++ themes/socialbase/src/Plugin/Preprocess/SocialbaseFieldset.php
@@ -2,27 +2,11 @@
 
 namespace Drupal\socialbase\Plugin\Preprocess;
 
-use Drupal\bootstrap\Plugin\Preprocess\PreprocessBase;
-use Drupal\bootstrap\Utility\Element;
-use Drupal\bootstrap\Utility\Variables;
-
 /**
  * Pre-processes variables for the "fieldset" theme hook.
  *
  * @ingroup plugins_preprocess
- *
- * @BootstrapPreprocess("fieldset")
+ * @deprecated
+ * @see \Drupal\socialbase\Plugin\Preprocess\Fieldset
  */
-class SocialbaseFieldset extends PreprocessBase {
-
-  /**
-   * {@inheritdoc}
-   */
-  protected function preprocessElement(Element $element, Variables $variables) {
-    if (isset($element['#type']) && $element['#type'] == ('radios' || 'checkboxes')) {
-      $variables['form_group'] = TRUE;
-    }
-
-  }
-
-}
+class SocialbaseFieldset extends Fieldset {}
